+++ Source/SwiftLintFramework/Rules/LineLengthRule.swift
@@ -23,7 +23,7 @@ public struct LineLengthRule: ParameterizedRule {
 
     public func validateFile(file: File) -> [StyleViolation] {
         return file.contents.lines().flatMap { line in
-            for parameter in self.parameters.reverse() {
+            for parameter in parameters.reverse() {
                 if line.content.characters.count > parameter.value {
                     return StyleViolation(type: .Length,
                         location: Location(file: file.path, line: line.index),