+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## Master
+
+##### Breaking
+
+* None.
+
+##### Enhancements
+
+* Configure SwiftLint via a YAML file:
+  Supports `disabled_rules`, `included` and `excluded`.
+  Pass a configuration file path to `--config`, defaults to `.swiftlint.yml`.  
+  [JP Simard](https://github.com/jpsim)
+  [#3](https://github.com/realm/SwiftLint/issues/3)
+
+##### Bug Fixes
+
+* None.
+
+
 ## 0.1.2: FabricSoftenerRule
 
 ##### Breaking